temporary folder/requirements.txt
asyncio
selenium
bs4
fastapi
pandas
uvicorn[standard]
openpyxl
mysql-connector-python
json